Server/backend/requirements.txt
django
PyMysql
wheel
mysqlclient
sorl-thumbnail
channels 
channels-redis
daphne
scipy
django-celery-results
python-decouple
Pillow
celery 
redis 
scipy
oauth2-provider
django-oauth-toolkit
oauthlib
requests-oauthlib
oauthlib[signedtoken]
pyjwt
oauthlib[rsa]
cryptography
django-oauth2-provider
djangorestframework
hypercorn
django-cors-headers
uvicorn
drf-yasg
drf-yasg[validation]
